add keep feature

+++ entrypoint.sh
@@ -10,16 +10,21 @@ OUT=""
 #loop for deleting old branches
 for i in $(git for-each-ref refs/remotes/origin --sort=committerdate  --format='%(HEAD)%(color:yellow)%(refname:short)%(color:reset) %(color:green)%(committerdate:raw)%(color:reset)')
 do
-    export IFS=$' '
-    elements=($i)
+    if [ $i -eq "origin/KEEP"* ]
+    then 
+      continue
+    else
+      export IFS=$' '
+      elements=($i)
     
-    if [ $TIME -gt ${elements[1]} ]
-    then
-        export IFS="/"
-	    inner_elements=(${elements[0]})
-        git push origin --delete ${inner_elements[1]}
-        #echo ${inner_elements[1]}
-        OUT="${OUT}, ${inner_elements[1]}"
+      if [ $TIME -gt ${elements[1]} ]
+      then
+          export IFS="/"
+	      inner_elements=(${elements[0]})
+          git push origin --delete ${inner_elements[1]}
+          #echo ${inner_elements[1]}
+          OUT="${OUT}, ${inner_elements[1]}"
+      fi
     fi
 done
 export IFS=$'\n'
